C structs
We have a structure in which there are the following fields: first name, last name and age. We must have two structure variables in the program, and the program must enter first name, last name, and age for both variables. Then the program should write which person is older and by how many years. Do this program 1. normally 2. with POINTER